Financial Connections: cleaned up tests to avoid redundant implementation of FinancialConnectionsAPIClient.
