+++ StripeFinancialConnections/StripeFinancialConnectionsTests/AccountFetcherTests.swift
@@ -10,7 +10,7 @@
 @testable import StripeFinancialConnections
 import XCTest
 
-class PaginatedAPIClient: FinancialConnectionsAPIClient {
+class PaginatedAPIClient: EmptyFinancialConnectionsAPIClient {
 
     // MARK: - Init
 
@@ -45,11 +45,7 @@ class PaginatedAPIClient: FinancialConnectionsAPIClient {
 
     // MARK: - FinancialConnectionsAPIClient
 
-    func generateSessionManifest(clientSecret: String, returnURL: String?) -> Promise<FinancialConnectionsSynchronize> {
-        return Promise<FinancialConnectionsSynchronize>()
-    }
-
-    func fetchFinancialConnectionsAccounts(
+    override func fetchFinancialConnectionsAccounts(
         clientSecret: String,
         startingAfterAccountId: String?
     ) -> Promise<StripeAPI.FinancialConnectionsSession.AccountList> {
@@ -64,83 +60,6 @@ class PaginatedAPIClient: FinancialConnectionsAPIClient {
         let list = StripeAPI.FinancialConnectionsSession.AccountList(data: subArray,
                                                hasMore: hasMore)
         return Promise<StripeAPI.FinancialConnectionsSession.AccountList>(value: list)
-    }
-
-    func fetchFinancialConnectionsSession(clientSecret: String) -> Promise<StripeAPI.FinancialConnectionsSession> {
-        return Promise<StripeAPI.FinancialConnectionsSession>()
-    }
-
-    func markConsentAcquired(clientSecret: String) -> Promise<FinancialConnectionsSessionManifest> {
-        return Promise<FinancialConnectionsSessionManifest>()
-    }
-
-    func fetchFeaturedInstitutions(clientSecret: String) -> Promise<FinancialConnectionsInstitutionList> {
-        return Promise<FinancialConnectionsInstitutionList>()
-    }
-
-    func fetchInstitutions(clientSecret: String, query: String) -> Promise<FinancialConnectionsInstitutionList> {
-        return Promise<FinancialConnectionsInstitutionList>()
-    }
-
-    func createAuthSession(clientSecret: String, institutionId: String) -> Promise<FinancialConnectionsAuthSession> {
-        return Promise<FinancialConnectionsAuthSession>()
-    }
-
-    func cancelAuthSession(clientSecret: String, authSessionId: String) -> Promise<FinancialConnectionsAuthSession> {
-        return Promise<FinancialConnectionsAuthSession>()
-    }
-
-    func fetchAuthSessionOAuthResults(clientSecret: String, authSessionId: String) -> Future<FinancialConnectionsMixedOAuthParams> {
-        return Promise<FinancialConnectionsMixedOAuthParams>()
-    }
-
-    func authorizeAuthSession(clientSecret: String, authSessionId: String, publicToken: String?) -> Promise<FinancialConnectionsAuthSession> {
-        return Promise<FinancialConnectionsAuthSession>()
-    }
-
-    func fetchAuthSessionAccounts(
-        clientSecret: String,
-        authSessionId: String,
-        initialPollDelay: TimeInterval
-    ) -> Future<FinancialConnectionsAuthSessionAccounts> {
-        return Promise<FinancialConnectionsAuthSessionAccounts>()
-    }
-
-    func selectAuthSessionAccounts(clientSecret: String, authSessionId: String, selectedAccountIds: [String]) -> Promise<FinancialConnectionsAuthSessionAccounts> {
-        return Promise<FinancialConnectionsAuthSessionAccounts>()
-    }
-
-    func markLinkingMoreAccounts(clientSecret: String) -> Promise<FinancialConnectionsSessionManifest> {
-        return Promise<FinancialConnectionsSessionManifest>()
-    }
-
-    func completeFinancialConnectionsSession(clientSecret: String) -> Future<StripeAPI.FinancialConnectionsSession> {
-        return Promise<StripeAPI.FinancialConnectionsSession>()
-    }
-
-    func attachBankAccountToLinkAccountSession(
-        clientSecret: String,
-        accountNumber: String,
-        routingNumber: String
-    ) -> Future<FinancialConnectionsPaymentAccountResource> {
-        return Promise<FinancialConnectionsPaymentAccountResource>()
-    }
-
-    func attachLinkedAccountIdToLinkAccountSession(
-        clientSecret: String,
-        linkedAccountId: String,
-        consumerSessionClientSecret: String?
-    ) -> Future<FinancialConnectionsPaymentAccountResource> {
-        return Promise<FinancialConnectionsPaymentAccountResource>()
-    }
-
-    func recordAuthSessionEvent(
-        clientSecret: String,
-        authSessionId: String,
-        eventNamespace: String,
-        eventName: String
-    ) -> Future<EmptyResponse> {
-        return Promise<EmptyResponse>()
     }
 
     // MARK: - Helpers
