+++ StripeFinancialConnections/StripeFinancialConnectionsTests/SessionFetcherTests.swift
@@ -10,7 +10,7 @@
 @testable import StripeFinancialConnections
 import XCTest
 
-class NoMoreAccountSessionAPIClient: FinancialConnectionsAPIClient {
+class NoMoreAccountSessionAPIClient: EmptyFinancialConnectionsAPIClient {
 
     // MARK: - Properties
 
@@ -24,17 +24,13 @@ class NoMoreAccountSessionAPIClient: FinancialConnectionsAPIClient {
 
     // MARK: - FinancialConnectionsAPIClient
 
-    func generateSessionManifest(clientSecret: String, returnURL: String?) -> Promise<FinancialConnectionsSynchronize> {
-        return Promise<FinancialConnectionsSynchronize>()
-    }
-
-    func fetchFinancialConnectionsAccounts(clientSecret: String, startingAfterAccountId: String?) -> Promise<StripeAPI.FinancialConnectionsSession.AccountList> {
+    override func fetchFinancialConnectionsAccounts(clientSecret: String, startingAfterAccountId: String?) -> Promise<StripeAPI.FinancialConnectionsSession.AccountList> {
         let account = StripeAPI.FinancialConnectionsAccount(balance: nil, balanceRefresh: nil, ownership: nil, ownershipRefresh: nil, displayName: nil, institutionName: "bank", last4: nil, category: .credit, created: 3, id: "12", livemode: false, permissions: nil, status: .active, subcategory: .checking, supportedPaymentMethodTypes: [.usBankAccount])
         let fullList = StripeAPI.FinancialConnectionsSession.AccountList(data: [account], hasMore: false)
         return Promise(value: fullList)
     }
 
-    func fetchFinancialConnectionsSession(clientSecret: String) -> Promise<StripeAPI.FinancialConnectionsSession> {
+    override func fetchFinancialConnectionsSession(clientSecret: String) -> Promise<StripeAPI.FinancialConnectionsSession> {
         let fullList = StripeAPI.FinancialConnectionsSession.AccountList(data: [], hasMore: hasMore)
         let sessionWithFullAccountList = StripeAPI.FinancialConnectionsSession(clientSecret: "las",
                                                                       id: "1234",
@@ -43,79 +39,6 @@ class NoMoreAccountSessionAPIClient: FinancialConnectionsAPIClient {
                                                                       paymentAccount: nil,
                                                                       bankAccountToken: nil)
         return Promise(value: sessionWithFullAccountList)
-    }
-
-    func markConsentAcquired(clientSecret: String) -> Promise<FinancialConnectionsSessionManifest> {
-        return Promise<FinancialConnectionsSessionManifest>()
-    }
-
-    func fetchFeaturedInstitutions(clientSecret: String) -> Promise<FinancialConnectionsInstitutionList> {
-        return Promise<FinancialConnectionsInstitutionList>()
-    }
-
-    func fetchInstitutions(clientSecret: String, query: String) -> Promise<FinancialConnectionsInstitutionList> {
-        return Promise<FinancialConnectionsInstitutionList>()
-    }
-
-    func createAuthSession(clientSecret: String, institutionId: String) -> Promise<FinancialConnectionsAuthSession> {
-        return Promise<FinancialConnectionsAuthSession>()
-    }
-
-    func cancelAuthSession(clientSecret: String, authSessionId: String) -> Promise<FinancialConnectionsAuthSession> {
-        return Promise<FinancialConnectionsAuthSession>()
-    }
-
-    func fetchAuthSessionOAuthResults(clientSecret: String, authSessionId: String) -> Future<FinancialConnectionsMixedOAuthParams> {
-        return Promise<FinancialConnectionsMixedOAuthParams>()
-    }
-
-    func authorizeAuthSession(clientSecret: String, authSessionId: String, publicToken: String?) -> Promise<FinancialConnectionsAuthSession> {
-        return Promise<FinancialConnectionsAuthSession>()
-    }
-
-    func fetchAuthSessionAccounts(
-        clientSecret: String,
-        authSessionId: String,
-        initialPollDelay: TimeInterval
-    ) -> Future<FinancialConnectionsAuthSessionAccounts> {
-        return Promise<FinancialConnectionsAuthSessionAccounts>()
-    }
-
-    func selectAuthSessionAccounts(clientSecret: String, authSessionId: String, selectedAccountIds: [String]) -> Promise<FinancialConnectionsAuthSessionAccounts> {
-        return Promise<FinancialConnectionsAuthSessionAccounts>()
-    }
-
-    func markLinkingMoreAccounts(clientSecret: String) -> Promise<FinancialConnectionsSessionManifest> {
-        return Promise<FinancialConnectionsSessionManifest>()
-    }
-
-    func completeFinancialConnectionsSession(clientSecret: String) -> Future<StripeAPI.FinancialConnectionsSession> {
-        return Promise<StripeAPI.FinancialConnectionsSession>()
-    }
-
-    func attachBankAccountToLinkAccountSession(
-        clientSecret: String,
-        accountNumber: String,
-        routingNumber: String
-    ) -> Future<FinancialConnectionsPaymentAccountResource> {
-        return Promise<FinancialConnectionsPaymentAccountResource>()
-    }
-
-    func attachLinkedAccountIdToLinkAccountSession(
-        clientSecret: String,
-        linkedAccountId: String,
-        consumerSessionClientSecret: String?
-    ) -> Future<FinancialConnectionsPaymentAccountResource> {
-        return Promise<FinancialConnectionsPaymentAccountResource>()
-    }
-
-    func recordAuthSessionEvent(
-        clientSecret: String,
-        authSessionId: String,
-        eventNamespace: String,
-        eventName: String
-    ) -> Future<EmptyResponse> {
-        return Promise<EmptyResponse>()
     }
 }
